Extract timer to separate component
